Introduce scheduling strategy for ParellelForWithWorkerId which schedules work items with a fixed shard size.
In comparison to ParallelForWithWorkerId, this method provides better scaling in some cases as the work is sharded into fewer parts.

Building on top of TransformRangeConcurrently, this method also passes a 'worker_id' or a 'thread_id' to the worker function, with the guarantee that the function calls with the same thread_id will not happen concurrently.

PiperOrigin-RevId: 264703742