Add op to name mapper for exporter
Add OpNameMapper interface that can be used to return a unique name for a given op or prefix. This class also allows setting the names for ops. Names returned are either the set or a unique name based on a given prefix (in some cases multiple ops can be and need to be associated with the same name). Adds two implementations of OpNameMapper, one that returns a name based on the location of the operation and the other that returns a short alphanumeric name.

There is probably some string mangling optimizations that could improve performance further here, but leaving it simpler until measured.

PiperOrigin-RevId: 265090266