Mock out multiprocessing.pool for multiprocessing.dummy.pool in unit tests.
multiprocessing is an inherently dangerous library due to its reliance on forking under the hood. However recent changes in TensorFlow interact particularly badly such that unit tests do not reliably run. In order to maintain some test coverage this change replaces the forkpool with a threadpool for unit tests, but exercises the rest of the code path.

In general users should switch to tf.data for input pipelines where concurrency is needed.

PiperOrigin-RevId: 265733412