Makes the tf.get_variable regularizers as lazy as possible.
Previous to this change, the values stored in the "REGULARIZATION_LOSSES" global collection were computed outside of any controllable context, making the behavior of graphs that computed the total loss within a control dependency unpredictable and counter-intuitive. Now the values stored in this collection by get_variable are "lazy tensors", that closely mimic the behavior and interface of regular tensors, but are computed (via the user provided callable) only when used in an expression (e.g. when added to the total loss). This fixes the aforementioned problem with control dependencies since now these regularization tensors will be computed in the local context where they are actually used.

PiperOrigin-RevId: 266353413